Fix agt CLI by moving finetune/convert/preprocess/predict implementations into the package
