+++ src/alphagenome_pytorch/extensions/inference/full_chromosome.py
@@ -39,6 +39,10 @@ from tqdm import tqdm
 from alphagenome_pytorch.genome import GenomeSequenceSource
 from alphagenome_pytorch.utils.sequence import sequence_to_onehot
 
+#: Chromosomes predicted when none are named: the main assembly, chr1..22 + chrX.
+#: Excludes chrY, chrM and scaffolds. Names not present in the FASTA are dropped.
+DEFAULT_CHROMOSOMES = [f"chr{i}" for i in range(1, 23)] + ["chrX"]
+
 # Lazy imports
 pyBigWig = None
 pyfaidx = None
@@ -501,7 +505,7 @@ def predict_full_chromosomes_to_bigwig(
 
     # Load genome
     if chromosomes is None:
-        chromosomes = [f"chr{i}" for i in range(1, 23)] + ["chrX"]
+        chromosomes = list(DEFAULT_CHROMOSOMES)
 
     genome = GenomeSequenceProvider(
         fasta_path,
@@ -624,7 +628,7 @@ def predict_full_chromosomes_to_anndata(
 
     config = config or TilingConfig()
     if chromosomes is None:
-        chromosomes = [f"chr{i}" for i in range(1, 23)] + ["chrX"]
+        chromosomes = list(DEFAULT_CHROMOSOMES)
 
     # `fasta_path` / `annotation_path` accept prebuilt objects too (handy for tests
     # and for reusing an already-loaded genome / annotation).
